+++ src/main/java/com/app/shwe/repository/ProcessRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.app.shwe.model.Process;
 
-public interface ProcessRepository extends JpaRepository<Process, Long> {
+public interface ProcessRepository extends JpaRepository<Process, Integer> {
 
 	List<Process> findAllByOrderByImageorderAsc();
